requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
groq==0.9.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
reportlab==4.0.9
Pillow==10.2.0
werkzeug==3.0.1
validate_email==1.3
flask-cors==4.0.0
